Calendar Addition and Discord Logo Change
After discussing with @JejuneKai on Discord we've decided it would be in
everyone's best interests to add a calendar button to the hotbar on the
subreddit.

We recently began planning events for the subreddit and found it
challenging to coordinate times with other tournaments (like PCL,
Pheonix, and Sylphin's). If were were able to have an icon that links to
a page on the wiki, we could consolidate a list of known events and
times. This way we could let others know when the subreddit's own events
would be and to plan around them (or possible allow tournaments to
submit their own dates to be added, I would handle if so desired).