News Ticker, FAQ, and Ad Fixes
So I rememberd some of the mods talking about not having enough stickies
for announcements/news on the subreddit (cause you only get two).

So, I figured why not free up some space while looking good at the same
time. Decided to add a news ticker. For reference it looks like this:

https://u.nya.is/kawvmw.mp4

Futhermore, I noticed that people were constantly asking the same stuff
on the subreddit over and over. We have a common requests link in the
sidebar but nobody ever reads it.

Added an FAQ section to the header:

https://u.nya.is/ybywgc.png

I talked with Jejune but we couldn't really come to the conclusion of
what to link to. I thought it would be nice to maybe actually update the
FAQ section on the subreddit, but that requires mod work. So we settled
on a megathread, that I already created here:
https://www.reddit.com/r/paragon/comments/69ncg0/common_requests_and_questions_thread/.

Lastly, I noticed if turn off adblock the ad that appears about the
stickies causes the green label to cut off and the ad background isn't
white, so I fixed that. Normally looks like this:

https://u.nya.is/jwuaop.png